# client/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Citation.is API",
    "version": "1.1.0",
    "description": "Public API for the citation.is open registry of verified scientific claims. Supports claim search, individual claim retrieval, real-time claim verification, entity lookup, and bulk data export. All endpoints are unauthenticated and rate-limited to 60 requests/minute.",
    "contact": {
      "name": "Citation.is",
      "url": "https://citation.is/developers"
    },
    "license": {
      "name": "CC BY 4.0",
      "url": "https://creativecommons.org/licenses/by/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://citation.is",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/public/verify-claim": {
      "post": {
        "operationId": "verifyClaim",
        "summary": "Verify a scientific claim",
        "description": "Verifies a single scientific claim in real-time against PubMed, UniProt, PubChem, and PDB. Returns a verdict, confidence score, rationale, and evidence URL. Rate-limited to 30 requests/minute.",
        "tags": ["Verification"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyClaimRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyClaimResponse" }
              }
            }
          },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VerifyClaimRequest": {
        "type": "object",
        "required": ["claim"],
        "properties": {
          "claim": { "type": "string", "description": "The scientific claim text to verify" },
          "vertical": {
            "type": "string",
            "enum": ["structural_biology", "salmon_biotech", "protein_supplement", "creatine_ergogenics", "gut_microbiome", "collagen_peptides", "plant_based_protein", "sports_nutrition_rct", "uniprot", "clinical_trials"],
            "description": "Optional: restrict verification to a specific research domain"
          }
        }
      },
      "VerifyClaimResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "claim": { "type": "string" },
          "verdict": { "type": "string", "enum": ["Supported", "Contradicted", "Insufficient Evidence", "Partially Supported"] },
          "rationale": { "type": "string" },
          "evidenceUrl": { "type": "string", "format": "uri" },
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "vertical": { "type": "string" },
          "apiVersion": { "type": "string" },
          "processedAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}